+++ src/main/java/faang/school/analytics/config/RedisConfiguration.java
@@ -1,6 +1,7 @@
 package faang.school.analytics.config;
 
 import faang.school.analytics.listener.AbstractEventListener;
+import faang.school.analytics.listener.LikeEventListener;
 import lombok.AllArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -15,6 +16,7 @@ import org.springframework.data.redis.listener.adapter.MessageListenerAdapter;
 public class RedisConfiguration {
     private final RedisProperties redisProperties;
     private final AbstractEventListener<?> mentorshipRequestsEventListener;
+    private final LikeEventListener likeEventListener;
 
     @Bean
     public JedisConnectionFactory redisConnectionFactory() {
@@ -39,7 +41,18 @@ public class RedisConfiguration {
         container.setConnectionFactory(redisConnectionFactory());
 
         container.addMessageListener(mentorshipRequestsListener(), mentorshipRequestsTopic());
+        container.addMessageListener(likeListener(likeEventListener), likeTopic());
 
         return container;
+    }
+
+    @Bean
+    ChannelTopic likeTopic(){
+        return new ChannelTopic(redisProperties.getChannel().getLike());
+    }
+
+    @Bean
+    MessageListenerAdapter likeListener(LikeEventListener likeEventListener){
+        return new MessageListenerAdapter(likeEventListener);
     }
 }
